move common funcs to catch2/common.hpp
